+++ code/src/sixsq/nuvla/server/resources/credential_infrastructure_service_kubernetes.clj
@@ -1,6 +1,6 @@
-(ns sixsq.nuvla.server.resources.credential-infrastructure-service-swarm
+(ns sixsq.nuvla.server.resources.credential-infrastructure-service-kubernetes
   "
-This resource contains the values necessary to access a Docker Swarm service.
+This resource contains the values necessary to access a Kubernetes service.
 These consist of a public 'cert' and the associated private 'key'. The
 certificate authority's public certificate, 'ca', should also be provided.
 "
@@ -10,7 +10,7 @@ certificate authority's public certificate, 'ca', should also be provided.
     [sixsq.nuvla.server.resources.common.std-crud :as std-crud]
     [sixsq.nuvla.server.resources.common.utils :as u]
     [sixsq.nuvla.server.resources.credential :as p]
-    [sixsq.nuvla.server.resources.credential-template-infrastructure-service-swarm :as tpl]
+    [sixsq.nuvla.server.resources.credential-template-infrastructure-service-kubernetes :as tpl]
     [sixsq.nuvla.server.resources.resource-metadata :as md]
     [sixsq.nuvla.server.resources.spec.credential-template-infrastructure-service-coe :as service-coe]
     [sixsq.nuvla.server.util.metadata :as gen-md]))